fixes + attempt to fix smoke bombs ( did not fix)

+++ code/game/objects/effects/effect_system.dm
@@ -223,8 +223,7 @@ steam.start() -- spawns the effect
 
 /obj/effect/effect/smoke/Initialize()
 	. = ..()
-	spawn(time_to_live)
-		fade_out()
+	addtimer(CALLBACK(src, PROC_REF(fade_out), time_to_live))
 
 
 /obj/effect/effect/smoke/Crossed(mob/living/carbon/M as mob )
@@ -257,10 +256,14 @@ steam.start() -- spawns the effect
 	fading = TRUE
 	frames = max(frames, 1) //We will just assume that by 0 frames, the coder meant "during one frame".
 	var/alpha_step = round(alpha / frames)
-	while(alpha > 0)
+	fade_step(alpha_step)
+
+/obj/effect/effect/smoke/proc/fade_step(alpha_step)
+	if(alpha > 0)
 		alpha = max(0, alpha - alpha_step)
-		sleep(world.tick_lag)
-	qdel(src)
+		addtimer(CALLBACK(src, PROC_REF(fade_step), alpha_step), 1 SECOND)
+	else
+		qdel(src)
 
 /////////////////////////////////////////////
 // Illumination
@@ -412,11 +415,10 @@ steam.start() -- spawns the effect
 			direction = pick(alldirs)
 		var/added_time = 1 SECOND
 		for(i=0, i<pick(0,1,1,1,2,2,2,3), i++)
-			addtimer(CALLBACK(src, .proc/move_smoke, smoke, direction), added_time)
+			addtimer(CALLBACK(src, PROC_REF(move_smoke), smoke, direction), added_time)
 			added_time += 1 SECOND
 
 /datum/effect/effect/system/smoke_spread/proc/move_smoke(atom/movable/smoke, move_dir)
-	SIGNAL_HANDLER
 	step(smoke, move_dir)
